Begin converting to TypeScript, rename .mjs to .ts
