+++ .eslintrc.json
@@ -3,11 +3,18 @@
         "node": true,
         "es2021": true
     },
+    "extends": [
+        "eslint:recommended",
+        "plugin:@typescript-eslint/recommended"
+    ],
+    "parser": "@typescript-eslint/parser",
     "parserOptions": {
         "ecmaVersion": "latest",
         "sourceType": "module"
     },
-    "extends": "eslint:recommended",
+    "plugins": [
+        "@typescript-eslint"
+    ],
     "rules": {
         "no-console": "error",
         "no-undef": "error"
