+++ .gitignore
@@ -1,5 +1,6 @@
 .idea
 eda-modbus-bridge.iml
 node_modules
+dist
 data
 *~
