+++ package.json
@@ -2,15 +2,16 @@
   "name": "eda-modbus-bridge",
   "version": "2.9.1",
   "description": "An HTTP/MQTT bridge for Enervent ventilation units with EDA automation",
-  "main": "eda-modbus-bridge.mjs",
+  "main": "eda-modbus-bridge.ts",
   "directories": {
     "doc": "docs"
   },
   "scripts": {
-    "test": "NODE_OPTIONS=--experimental-vm-modules jest",
-    "check-prettier": "prettier --check eda-modbus-bridge.mjs app/ tests/",
-    "prettier": "prettier --write eda-modbus-bridge.mjs app/ tests/",
-    "lint": "eslint eda-modbus-bridge.mjs 'app/**/*.mjs'"
+    "build": "tsc",
+    "test": "jest",
+    "check-prettier": "prettier --check eda-modbus-bridge.ts app/ tests/",
+    "prettier": "prettier --write eda-modbus-bridge.ts app/ tests/",
+    "lint": "eslint eda-modbus-bridge.ts 'app/**/*.ts'"
   },
   "repository": {
     "type": "git",
@@ -36,11 +37,17 @@
     "yargs": "^16.2.0"
   },
   "devDependencies": {
+    "@tsconfig/node14": "^14.1.2",
+    "@types/jest": "^29.5.12",
+    "@typescript-eslint/eslint-plugin": "^8.1.0",
+    "@typescript-eslint/parser": "^8.1.0",
     "eslint": "^8.47.0",
     "eslint-plugin-import": "^2.28.0",
     "eslint-plugin-n": "^16.0.1",
     "eslint-plugin-promise": "^6.1.1",
-    "jest": "^27.4.7",
-    "prettier": "^2.5.1"
+    "jest": "^29.7.0",
+    "prettier": "^2.5.1",
+    "ts-jest": "^29.2.4",
+    "typescript": "^5.5.4"
   }
 }
