+++ systemd/eda-modbus-bridge.service
@@ -3,7 +3,7 @@ Description=HTTP bridge for Enervent ventilation units with EDA automation
 
 [Service]
 Type=simple
-ExecStart=/usr/bin/node /opt/eda-modbus-bridge/eda-modbus-bridge.mjs -d /dev/ttyUSB0
+ExecStart=/usr/bin/node /opt/eda-modbus-bridge/dist/eda-modbus-bridge.js -d /dev/ttyUSB0
 Restart=always
 RestartSec=5
